requirements.txt
google-cloud-bigquery==3.4.2
boxsdk==3.6.2